+++ src/features/profile/components/Tabs.tsx
@@ -8,6 +8,7 @@ import { cn } from '@/lib/classnames'
 const tabs = [
   { name: 'Created Builds', href: '/profile/created-builds' },
   { name: 'Favorited Builds', href: '/profile/favorited-builds' },
+  { name: 'Loadout Builds', href: '/profile/loadout-builds' },
 ]
 
 export function Tabs() {